requirements.txt
google-generativeai==0.8.6
python-telegram-bot==22.7
ddgs>=9.0.0
requests>=2.32.0
beautifulsoup4>=4.12.0
